+++ include/graph/impl/SessionLocalStorage.cpp
@@ -1,0 +1,112 @@
+//
+// @author raver119@gmail.com
+//
+
+#include <VariableSpace.h>
+#include <Stash.h>
+#include <graph/SessionLocalStorage.h>
+
+namespace nd4j {
+    namespace graph {
+
+        template <typename T>
+        SessionLocalStorage<T>::SessionLocalStorage(VariableSpace<T>* variableSpace, Stash<T>* stash) {
+            // we start from 1, since key 0 holds original VariableSpace
+            _sessionCounter.store(1);
+            _variableSpace = variableSpace;
+            _stash = stash;
+        }
+
+        template <typename T>
+        VariableSpace<T>* SessionLocalStorage<T>::localVariableSpace(Nd4jIndex sessionId) {
+            _mutex.lock();
+            auto varSpace = _threadVariableSpace.at(sessionId);
+            _mutex.unlock();
+
+            return varSpace;
+        }
+
+        template <typename T>
+        VariableSpace<T>* SessionLocalStorage<T>::localVariableSpace() {
+            return localVariableSpace(getSessionId());
+        }
+
+
+        template <typename T>
+        Nd4jIndex SessionLocalStorage<T>::getThreadId() {
+#ifdef __APPLE__
+            // syscall?
+#elif _WIN32
+            // some win32api
+#else
+    // syscall!
+#endif
+            auto id=std::this_thread::get_id();
+            uint64_t* ptr=(uint64_t*) &id;
+            return (*ptr);
+        }
+
+        template <typename T>
+        int SessionLocalStorage<T>::numberOfSessions() {
+            _mutex.lock();
+            int size = (int) _threadSession.size();
+            _mutex.unlock();
+            return size;
+        }
+
+        template <typename T>
+        void SessionLocalStorage<T>::endSession(Nd4jIndex sessionId) {
+            // we should delete specific holders here
+            _mutex.lock();
+            _threadVariableSpace.erase(sessionId);
+            _mutex.unlock();
+        }
+
+        template <typename T>
+        void SessionLocalStorage<T>::endSession() {
+            auto tid = getThreadId();
+
+            _mutex.lock();
+
+            auto ntid = _threadSession[tid];
+            _threadSession.erase(tid);
+
+            _mutex.unlock();
+
+            endSession(ntid);
+        }
+
+        template <typename T>
+        Nd4jIndex SessionLocalStorage<T>::getSessionId() {
+            auto tid = getThreadId();
+
+            _mutex.lock();
+            auto ntid = _threadSession[tid];
+
+            _mutex.unlock();
+
+            return ntid;
+        }
+
+        template <typename T>
+        Nd4jIndex nd4j::graph::SessionLocalStorage<T>::startSession() {
+            auto tid = getThreadId();
+
+            nd4j_debug("Adding ThreadId: %i;\n", (int) tid);
+            Nd4jIndex ntid = _sessionCounter++;
+            _mutex.lock();
+
+            _threadSession[tid] = ntid;
+            _threadVariableSpace[ntid] = _variableSpace->clone();
+
+            _mutex.unlock();
+
+            return ntid;
+        }
+
+        template class SessionLocalStorage<float>;
+        template class SessionLocalStorage<float16>;
+        template class SessionLocalStorage<double>;
+    }
+}
+
