+++ include/indexing/IndicesList.h
@@ -13,27 +13,12 @@ namespace nd4j {
     protected:
         std::vector<NDIndex *> _indices;
     public:
-        IndicesList(std::initializer_list<NDIndex *> list) {
-            for (auto v: list)
-                _indices.push_back(v);
-        }
+        IndicesList(std::initializer_list<NDIndex *> list);
 
         int size();
         NDIndex* at(int idx);
 
-        ~IndicesList() {
-            for(auto v: _indices)
-                delete v;
-        }
+        ~IndicesList();
     };
 }
-
-int nd4j::IndicesList::size() {
-    return (int) _indices.size();
-}
-
-nd4j::NDIndex* nd4j::IndicesList::at(int idx) {
-    return _indices.at(idx);
-}
-
 #endif //LIBND4J_INDICESLIST_H
