+++ tests/layers_tests/CMakeLists.txt
@@ -58,7 +58,7 @@ ENDIF(${CMAKE_SYSTEM_NAME} MATCHES "Linux")
 
 message("CPU BLAS")
 add_definitions(-D__CPUBLAS__=true)
-add_library(nd4jcpu SHARED ../../blas/cpu/NativeOps.cpp ../../include/cnpy/cnpy.cpp  ../../include/nd4jmemset.h ../../include/nd4jmalloc.h ../../include/ops/declarable/generic/parity_ops.h ../../include/helpers/helper_hash.h ../../include/memory/cpu/Workspace.cpp ../../include/ops/declarable/generic/helpers/convolutions.h ../../include/graph/ShapeList.h ../../include/ops/declarable/generic/convo/convo_ops.h ../../include/helpers/ArrayUtils.h ../../include/indexing/NDIndex.h ../../include/indexing/IndicesList.h ../../include/graph/SessionLocalStorage.h ../../include/graph/Stash.h ../../blas/Environment.cpp)
+add_library(nd4jcpu SHARED ../../blas/cpu/NativeOps.cpp ../../include/cnpy/cnpy.cpp  ../../include/nd4jmemset.h ../../include/nd4jmalloc.h ../../include/ops/declarable/generic/parity_ops.h ../../include/helpers/helper_hash.h ../../include/memory/cpu/Workspace.cpp ../../include/ops/declarable/generic/helpers/convolutions.h ../../include/graph/ShapeList.h ../../include/ops/declarable/generic/convo/convo_ops.h ../../include/helpers/ArrayUtils.h ../../include/graph/SessionLocalStorage.h  ../../blas/Environment.cpp ../../include/graph/impl/Stash.cpp ../../include/indexing/impl/NDIndex.cpp ../../include/indexing/impl/IndicesList.cpp ../../include/graph/impl/Variable.cpp ../../include/graph/impl/VariableSpace.cpp ../../include/graph/impl/ShapeList.cpp ../../include/graph/impl/SessionLocalStorage.cpp ../../include/graph/impl/Node.cpp ../../include/graph/impl/Graph.cpp ../../include/graph/impl/Block.cpp ../../include/graph/impl/ArrayList.cpp)
 
 if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU" AND "${CMAKE_CXX_COMPILER_VERSION}" VERSION_LESS 4.9)
     message(FATAL_ERROR "You need at least GCC 4.9")
